fix changelog notification
