+++ src/kmc-app/components/app-menu/app-menu.component.ts
@@ -11,6 +11,7 @@ import { KmcLoggerConfigurator } from 'app-shared/kmc-shell/kmc-logs/kmc-logger-
 import { KalturaLogger } from '@kaltura-ng/kaltura-logger/kaltura-logger.service';
 import { KMCAppMenuItem, KmcMainViewsService } from 'app-shared/kmc-shared/kmc-views';
 import { ContextualHelpLink, ContextualHelpService } from 'app-shared/kmc-shared/contextual-help/contextual-help.service';
+import { globalConfig } from 'config/global';
 
 @Component({
     selector: 'kKMCAppMenu',
@@ -24,6 +25,8 @@ import { ContextualHelpLink, ContextualHelpService } from 'app-shared/kmc-shared
 export class AppMenuComponent implements OnInit, OnDestroy{
 
     @ViewChild('helpmenu') private _helpmenu: PopupWidgetComponent;
+    private _appCachedVersionToken = 'kmc-cached-app-version';
+
     public _showChangelog = false;
     public _helpMenuOpened = false;
     public _powerUser = false;
@@ -74,6 +77,8 @@ export class AppMenuComponent implements OnInit, OnDestroy{
     }
 
     ngOnInit() {
+        const cachedVersion = this._browserService.getFromLocalStorage(this._appCachedVersionToken);
+        this._showChangelog = cachedVersion !== globalConfig.client.appVersion;
     }
 
     setSelectedRoute(path) {
@@ -112,5 +117,10 @@ export class AppMenuComponent implements OnInit, OnDestroy{
 
 
     ngOnDestroy() {
+    }
+
+    public _changelogPopupOpened(): void {
+        this._showChangelog = false;
+        this._browserService.setInLocalStorage(this._appCachedVersionToken, globalConfig.client.appVersion);
     }
 }
